Re-organize more modules to avoid expensive import costs.
