+++ st2common/st2common/script_setup.py
@@ -27,8 +27,8 @@ import logging as stdlib_logging
 from oslo_config import cfg
 
 from st2common import log as logging
-from st2common.service_setup import db_setup
-from st2common.service_setup import db_teardown
+from st2common.database_setup import db_setup
+from st2common.database_setup import db_teardown
 from st2common.logging.filters import LogLevelFilter
 from st2common.transport.bootstrap_utils import register_exchanges_with_retry
